Option-2_Container/requirements.txt
ibm-cos-sdk
ibm-cloud-sdk-core
python-dotenv
tqdm